requirements.lock.txt
aenum==3.1.16
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
aniso8601==10.0.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
anywidget==0.9.21
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
bqplot==0.12.45
branca==0.8.2
cachelib==0.13.0
cachetools==7.0.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-plugins==1.1.1.2
cligj==0.7.2
color-operations==0.2.0
comm==0.2.3
contextily==1.7.0
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.2.2
cycler==0.12.1
datasets==4.6.1
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
docstring_parser==0.17.0
duckdb==1.4.4
einops==0.8.2
eval_type_backport==0.3.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
fiona==1.10.1
Flask==3.1.3
Flask-Caching==2.3.1
flask-cors==6.0.2
flask-restx==1.3.2
folium==0.20.0
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.12.0
gdown==5.2.1
geoai-py==0.32.0
geographiclib==2.1
geojson==3.2.0
geopandas==1.1.2
geopy==2.4.1
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
hydra-core==1.3.2
idna==3.11
ImageIO==2.37.2
importlib_resources==6.5.2
ipyevents==2.0.4
ipyfilechooser==0.6.0
ipyleaflet==0.20.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
ipytree==0.2.2
ipyvue==1.12.0
ipyvuetify==1.11.3
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonargparse==4.46.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-leaflet==0.20.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_proxy==4.4.0
jupyter_server_terminals==0.5.4
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
kornia==0.8.2
kornia_rs==0.1.10
lark==1.3.1
lazy_loader==0.4
leafmap==0.60.2
lightly==1.5.22
lightly-utils==0.0.2
lightning==2.6.1
lightning-utilities==0.15.3
localtileserver==0.11.0
mapclassify==2.10.0
maplibre==0.3.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mercantile==1.2.1
mistune==3.2.0
morecantile==7.0.3
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
narwhals==2.17.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
networkx==3.6.1
numexpr==2.14.1
numpy==2.3.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
opencv-python-headless==4.13.0.92
osmnx==2.1.0
overturemaps==0.19.0
packaging==26.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.6
pexpect==4.9.0
pillow==12.0.0
planetary-computer==1.0.0
platformdirs==4.9.2
plotly==6.6.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
psygnal==0.15.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pycparser==3.0
pycocotools==2.0.10
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyogrio==0.12.1
pyparsing==3.3.2
pyproj==3.7.2
PySocks==1.7.1
pystac==1.14.3
pystac-client==0.9.0
python-box==7.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.0.0
pytorch-lightning==2.6.1
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
rasterio==1.5.0
rasterstats==0.20.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rio-cogeo==7.0.1
rio-tiler==8.0.5
rioxarray==0.21.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
scooby==0.11.0
seaborn==0.13.2
segmentation_models_pytorch==0.5.0
Send2Trash==2.1.0
server-thread==0.3.0
setuptools==70.2.0
shapely==2.1.2
shellingham==1.5.4
simpervisor==1.0.0
simplejson==3.20.2
simsimd==6.5.15
six==1.17.0
smoothify==0.2.1
soupsieve==2.8.3
stack-data==0.6.3
stringzilla==4.6.0
sympy==1.14.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2026.3.3
timm==1.0.25
tinycss2==1.4.0
tokenizers==0.22.2
torch==2.10.0+cu126
torchgeo==0.9.0
torchinfo==1.8.0
torchmetrics==1.8.2
torchvision==0.25.0+cu126
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
traittypes==0.2.3
transformers==5.2.0
triton==3.6.0
typer==0.24.1
typer-slim==0.24.0
typeshed_client==2.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
uvicorn==0.41.0
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.6
whitebox==2.3.6
whiteboxgui==2.3.0
widgetsnbextension==4.0.15
xarray==2025.11.0
xxhash==3.6.0
xyzservices==2025.11.0
yarl==1.23.0
